+++ src/main/java/de/tum/in/ase/staticCodeAnalysisParser/domain/StaticCodeAnalysisTool.java
@@ -1,0 +1,5 @@
+package de.tum.in.ase.staticCodeAnalysisParser.domain;
+
+public enum StaticCodeAnalysisTool {
+    SPOTBUGS, CHECKSTYLE, PMD
+}
