+++ src/main/java/de/tum/in/ase/staticCodeAnalysisParser/exception/ParserException.java
@@ -1,0 +1,8 @@
+package de.tum.in.ase.staticCodeAnalysisParser.exception;
+
+public class ParserException extends Exception {
+
+    public ParserException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
