+++ src/main/java/de/tum/in/ase/staticCodeAnalysisParser/exception/UnsupportedToolException.java
@@ -1,0 +1,8 @@
+package de.tum.in.ase.staticCodeAnalysisParser.exception;
+
+public class UnsupportedToolException extends RuntimeException {
+
+    public UnsupportedToolException(String message) {
+        super(message);
+    }
+}
