--- src/main/java/de/tum/in/www1/staticCodeAnalysisParser/domain/StaticCodeAnalysisTool.java
@@ -1,5 +1,0 @@
-package de.tum.in.www1.staticCodeAnalysisParser.domain;
-
-public enum StaticCodeAnalysisTool {
-    SPOTBUGS, CHECKSTYLE, PMD
-}
