--- src/main/java/de/tum/in/www1/staticCodeAnalysisParser/exception/ParserException.java
@@ -1,8 +1,0 @@
-package de.tum.in.www1.staticCodeAnalysisParser.exception;
-
-public class ParserException extends Exception {
-
-    public ParserException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
