--- src/main/java/de/tum/in/www1/staticCodeAnalysisParser/exception/UnsupportedToolException.java
@@ -1,8 +1,0 @@
-package de.tum.in.www1.staticCodeAnalysisParser.exception;
-
-public class UnsupportedToolException extends RuntimeException {
-
-    public UnsupportedToolException(String message) {
-        super(message);
-    }
-}
