+++ src/test/java/de/tum/in/ase/staticCodeAnalysisParser/AppTest.java
@@ -1,0 +1,20 @@
+package de.tum.in.ase.staticCodeAnalysisParser;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest 
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        assertTrue( true );
+    }
+}
