--- src/test/java/de/tum/in/www1/staticCodeAnalysisParser/AppTest.java
@@ -1,20 +1,0 @@
-package de.tum.in.www1.staticCodeAnalysisParser;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import org.junit.jupiter.api.Test;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
